# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ledgerway procurement-to-payment API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orgs/{orgSlug}/requisitions:
    get:
      operationId: listRequisitions
      tags: [requisitions]
      summary: List requisitions
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
        - $ref: "#/components/parameters/Status"
        - $ref: "#/components/parameters/Search"
      responses:
        "200":
          description: Requisition list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Requisition"
    post:
      operationId: createRequisition
      tags: [requisitions]
      summary: Create a requisition
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequisitionInput"
      responses:
        "201":
          description: Created requisition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Requisition"
  /orgs/{orgSlug}/requisitions/{requisitionId}:
    get:
      operationId: getRequisition
      tags: [requisitions]
      summary: Get a requisition
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
        - $ref: "#/components/parameters/RequisitionId"
      responses:
        "200":
          description: Requisition detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequisitionDetail"
        "404":
          $ref: "#/components/responses/NotFound"
  /orgs/{orgSlug}/requisitions/{requisitionId}/submit:
    post:
      operationId: submitRequisition
      tags: [requisitions]
      summary: Submit a requisition for approval
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
        - $ref: "#/components/parameters/RequisitionId"
      responses:
        "200":
          description: Submitted requisition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequisitionDetail"
  /orgs/{orgSlug}/requisitions/{requisitionId}/withdraw:
    post:
      operationId: withdrawRequisition
      tags: [requisitions]
      summary: Withdraw a pending requisition
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
        - $ref: "#/components/parameters/RequisitionId"
      responses:
        "200":
          description: Withdrawn requisition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequisitionDetail"
...
  /orgs/{orgSlug}/purchase-orders:
    get:
      operationId: listPurchaseOrders
      tags: [purchase-orders]
      summary: List purchase orders
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
        - $ref: "#/components/parameters/Status"
        - $ref: "#/components/parameters/Search"
      responses:
        "200":
          description: Purchase order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseOrder"
    post:
      operationId: createPurchaseOrder
      tags: [purchase-orders]
      summary: Create a purchase order
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrderInput"
      responses:
        "201":
          description: Created purchase order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /orgs/{orgSlug}/purchase-orders/{purchaseOrderId}/issue:
    post:
      operationId: issuePurchaseOrder
      tags: [purchase-orders]
      summary: Issue a purchase order and commit budget
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
        - name: purchaseOrderId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Issued purchase order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /orgs/{orgSlug}/vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List vendors
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
        - $ref: "#/components/parameters/Search"
      responses:
        "200":
          description: Vendor list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Create a vendor
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Created vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /orgs/{orgSlug}/receiving:
    post:
      operationId: createReceivingRecord
      tags: [receiving]
      summary: Record received goods or services
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceivingInput"
      responses:
        "201":
          description: Receiving record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceivingRecord"
...
components:
  parameters:
    OrgSlug:
      name: orgSlug
      in: path
      required: true
      schema: { type: string, minLength: 1 }
    RequisitionId:
      name: requisitionId
      in: path
      required: true
      schema: { type: string, format: uuid }
    Status:
      name: status
      in: query
      required: false
      schema: { type: string }
    Search:
      name: search
      in: query
      required: false
      schema: { type: string }
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    RequisitionInput:
      type: object
      required: [description, estimatedAmountCents, costCenter, category, neededByDate, justification]
      properties:
        description: { type: string, minLength: 3 }
        estimatedAmountCents: { type: integer, minimum: 1 }
        costCenter: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        neededByDate: { type: string, format: date }
        justification: { type: string, minLength: 10 }
        vendorName: { type: ["string", "null"] }
    Requisition:
      type: object
      required: [id, number, description, estimatedAmountCents, currency, costCenter, category, requester, neededByDate, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        number: { type: string }
        description: { type: string }
        estimatedAmountCents: { type: integer }
        currency: { type: string }
        costCenter: { type: string }
        category: { type: string }
        requester: { type: string }
        neededByDate: { type: string, format: date }
        status: { type: string }
        createdAt: { type: string, format: date-time }
        currentApprover: { type: ["string", "null"] }
        justification: { type: ["string", "null"] }
    RequisitionDetail:
      allOf:
        - $ref: "#/components/schemas/Requisition"
        - type: object
          required: [approvalSteps]
          properties:
            approvalSteps:
              type: array
              items: { $ref: "#/components/schemas/ApprovalStep" }
    ApprovalStep:
      type: object
      required: [id, order, approver, status]
      properties:
        id: { type: string, format: uuid }
        order: { type: integer }
        approver: { type: string }
        status: { type: string }
        decidedAt: { type: ["string", "null"], format: date-time }
        note: { type: ["string", "null"] }
...
    PurchaseOrderInput:
      type: object
      required: [vendorName, costCenter, lineItems]
      properties:
        requisitionId: { type: ["string", "null"], format: uuid }
        vendorName: { type: string, minLength: 1 }
        costCenter: { type: string }
        lineItems:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/PurchaseOrderLineInput" }
    PurchaseOrderLineInput:
      type: object
      required: [description, quantity, unitPriceCents]
      properties:
        description: { type: string }
        quantity: { type: number, minimum: 0.01 }
        unitPriceCents: { type: integer, minimum: 1 }
        glCode: { type: ["string", "null"] }
    PurchaseOrder:
      type: object
      required: [id, number, vendorName, totalAmountCents, currency, status, issuedAt, lineItems]
      properties:
        id: { type: string, format: uuid }
        number: { type: string }
        vendorName: { type: string }
        totalAmountCents: { type: integer }
        currency: { type: string }
        status: { type: string }
        issuedAt: { type: ["string", "null"], format: date-time }
        lineItems:
          type: array
          items: { $ref: "#/components/schemas/PurchaseOrderLineInput" }
    VendorInput:
      type: object
      required: [name, paymentTerms]
      properties:
        name: { type: string, minLength: 2 }
        paymentTerms: { type: string }
        preferred: { type: boolean }
    Vendor:
      type: object
      required: [id, name, status, paymentTerms, spendCents, exceptionRate, onTimeRate]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        status: { type: string }
        paymentTerms: { type: string }
        spendCents: { type: integer }
        exceptionRate: { type: number }
        onTimeRate: { type: number }
    ReceivingInput:
      type: object
      required: [purchaseOrderId, quantity, conditionNotes]
      properties:
        purchaseOrderId: { type: string, format: uuid }
        quantity: { type: number, minimum: 0.01 }
        conditionNotes: { type: string }
    ReceivingRecord:
      type: object
      required: [id, purchaseOrderId, quantity, receivedAt, conditionNotes]
      properties:
        id: { type: string, format: uuid }
        purchaseOrderId: { type: string, format: uuid }
        quantity: { type: number }
        receivedAt: { type: string, format: date-time }
        conditionNotes: { type: string }
...